fix(ses): SendDataPoint counters are int64 on the wire
The kind-mismatch sweep's XML low-confidence bucket (gopherstack-z5je9)
found one real bug in 23 hits across 18 modules: ses SendDataPoint's
Bounces/Complaints/DeliveryAttempts/Rejects were float64 while the real
type is int64 (ses@v1.37.4 types/types.go:1208). encoding/xml switches
float64 to scientific notation at 1e6, which the SDK's strconv.ParseInt
deserializer rejects. Both structs are int64 now; a raw-body real-client
test proves the old shape fails to decode at 1e6 and the new one decodes at
any magnitude. Recomputed per call; no persisted change.

ec2query was analysed read-only (180 exact-name types 0 hits, 16 key-join
hits all false positives, 10 float fields and ~60 timestamp sites checked)
with zero bugs; the technique has low recall on ec2 because its wire DTOs
rarely share AWS struct names. Remaining false positives were sibling
wrapper-type name collisions and literal "true"/"false" strings.

Closes: gopherstack-z5je9

+++ services/ses/handler_sending_stats.go
@@ -73,11 +73,11 @@ type getSendQuotaResponse struct {
 }
 
 type xmlSendDataPoint struct {
-	Timestamp        string  `xml:"Timestamp"`
-	DeliveryAttempts float64 `xml:"DeliveryAttempts"`
-	Bounces          float64 `xml:"Bounces"`
-	Complaints       float64 `xml:"Complaints"`
-	Rejects          float64 `xml:"Rejects"`
+	Timestamp        string `xml:"Timestamp"`
+	DeliveryAttempts int64  `xml:"DeliveryAttempts"`
+	Bounces          int64  `xml:"Bounces"`
+	Complaints       int64  `xml:"Complaints"`
+	Rejects          int64  `xml:"Rejects"`
 }
 
 type xmlSendDataPointList struct {
